Support to hide selected nodes
This patch supports to hide selected nodes.

Namely, texts on selected (i.e. clicked) nodes become red from black,
and a subsequent keyboard action of pressing the 'd' key removes the
nodes with related links.

Pressing the escape key displays hidden nodes and links back to the
screen.

Note that this feature is disabled when tooltip: 'click' is set and
enabled only when tooltip: 'hover' is set.